+++ src/main/java/com/swyp3/skin/admin/product/controller/AdminProductController.java
@@ -1,19 +1,13 @@
 package com.swyp3.skin.admin.product.controller;
 
+import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+@Controller
 @RequestMapping("admin")
-public class AdminController {
-
-    @GetMapping
-    public String index(Model model) {
-        model.addAttribute("pageTitle", "Admin Dashboard");
-        model.addAttribute("systemName", "Skin Service");
-        model.addAttribute("userRole", "ADMIN");
-        return "admin/index";
-    }
+public class AdminProductController {
 
     @GetMapping("/products")
     public String products(Model model) {
