+++ src/main/java/com/swyp3/skin/domain/product/domain/enums/ProductCategory.java
@@ -1,3 +1,3 @@
 package com.swyp3.skin.domain.product.domain.enums;
 
-public enum ProductCategory { TONER, SERUM, CREAM, SUNSCREEN }
+public enum ProductCategory { CLEANSER, TONER, ESSENCE, SERUM, AMPOULE, LOTION, CREAM, SUNSCREEN, MASK, SPOT }
